requirements.txt
ghapi==0.1.21
GitPython==3.1.24
yt-dlp==2022.01.08
